update network for wsl
Rename WSL classes, update learning rate scheduler and update dual-branch network
