+++ pymic/net_run_wsl/wsl_abstract.py
@@ -5,7 +5,7 @@ from pymic.net_run.agent_seg import SegmentationAgent
 
 class WSLSegAgent(SegmentationAgent):
     """
-    Training and testing agent for semi-supervised segmentation
+    Training and testing agent for weakly supervised segmentation
     """
     def __init__(self, config, stage = 'train'):
         super(WSLSegAgent, self).__init__(config, stage)
@@ -13,7 +13,7 @@ class WSLSegAgent(SegmentationAgent):
     def training(self):
         pass
         
-    def write_scalars(self, train_scalars, valid_scalars, glob_it):
+    def write_scalars(self, train_scalars, valid_scalars, lr_value, glob_it):
         loss_scalar ={'train':train_scalars['loss'], 
                       'valid':valid_scalars['loss']}
         loss_sup_scalar  = {'train':train_scalars['loss_sup']}
@@ -23,6 +23,7 @@ class WSLSegAgent(SegmentationAgent):
         self.summ_writer.add_scalars('loss_sup', loss_sup_scalar, glob_it)
         self.summ_writer.add_scalars('loss_reg', loss_upsup_scalar, glob_it)
         self.summ_writer.add_scalars('regular_w', {'regular_w':train_scalars['regular_w']}, glob_it)
+        self.summ_writer.add_scalars('lr', {"lr": lr_value}, glob_it)
         self.summ_writer.add_scalars('dice', dice_scalar, glob_it)
         class_num = self.config['network']['class_num']
         for c in range(class_num):
