+++ pymic/net_run_wsl/wsl_gatedcrf.py
@@ -86,7 +86,8 @@ class WSLGatedCRF(WSLSegAgent):
 
             loss.backward()
             self.optimizer.step()
-            self.scheduler.step()
+            if(not keyword_match(self.config['training']['lr_scheduler'], "ReduceLROnPlateau")):
+                self.scheduler.step()
 
             train_loss = train_loss + loss.item()
             train_loss_sup = train_loss_sup + loss_sup.item()
